e2e: serial: split config-related tests
Split off tests which verify how NROP reacts to config
changes. These these tests don't depend on the TM scope,
so we de-entangle them from the `scope=container` constraint.
